Rename tapWithOffset and clickWithOffset.
