+++ src/e2e/iOS/helpers/tap.ts
@@ -13,9 +13,9 @@ interface Options {
 }
 
 /**
- * Tap the given node with offset.
+ * Tap a node with an optional text offset or x,y offset.
  */
-const tapWithOffset = async (browser: Browser<'async'>, nodeHandle: Element<'async'>, { horizontalTapLine = 'left', offset, x = 0, y = 0 }: Options = {}) => {
+const tap = async (browser: Browser<'async'>, nodeHandle: Element<'async'>, { horizontalTapLine = 'left', offset, x = 0, y = 0 }: Options = {}) => {
   const boundingBox = await browser.getElementRect(nodeHandle.elementId)
   if (!boundingBox) throw new Error('Bouding box of editable not found.')
 
@@ -61,4 +61,4 @@ const tapWithOffset = async (browser: Browser<'async'>, nodeHandle: Element<'asy
 
 }
 
-export default tapWithOffset
+export default tap
